+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
  * Update the name of the package from `logrus_logstash` to `logrustash`
  * Add TimeFormat to Hook
- * Replace the old logrus package path: `github.com/Sirupsen/logrus` with `github.com/sirupsen/logrus` 
+ * Replace the old logrus package path: `github.com/EliteX23/logrus` with `github.com/EliteX23/logrus` 
 
 ## 0.3